Remove the isEmptyElement util from adapter-utils
