+++ packages/workspace/src/workspace/nacl_files/multi_env/routers.ts
@@ -13,11 +13,11 @@
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
-import { getChangeElement, ElemID, Value, DetailedChange, ChangeDataType, Element } from '@salto-io/adapter-api'
+import { getChangeElement, ElemID, Value, DetailedChange, ChangeDataType, Element, isObjectType, isPrimitiveType, isInstanceElement, isField } from '@salto-io/adapter-api'
 import _ from 'lodash'
 import path from 'path'
 import { promises, values } from '@salto-io/lowerdash'
-import { resolvePath, filterByID, detailedCompare, applyFunctionToChangeData, isEmptyElement } from '@salto-io/adapter-utils'
+import { resolvePath, filterByID, detailedCompare, applyFunctionToChangeData } from '@salto-io/adapter-utils'
 import { ElementsSource } from '../../elements_source'
 import {
   projectChange, projectElementOrValueToEnv, createAddChange, createRemoveChange,
@@ -48,11 +48,30 @@ const toPathHint = (filename: string): string[] => {
   return [...dirPathSplitted, path.basename(filename, path.extname(filename))]
 }
 
+const isEmptyAnnoAndAnnoTypes = (element: Element): boolean =>
+  (_.isEmpty(element.annotations) && _.isEmpty(element.annotationTypes))
+
+const isEmptyChangeElement = (element: Element): boolean => {
+  if (isObjectType(element)) {
+    return isEmptyAnnoAndAnnoTypes(element) && _.isEmpty(element.fields)
+  }
+  if (isPrimitiveType(element)) {
+    return isEmptyAnnoAndAnnoTypes(element)
+  }
+  if (isInstanceElement(element)) {
+    return _.isEmpty(element.annotations) && _.isEmpty(element.value)
+  }
+  if (isField(element)) {
+    return _.isEmpty(element.annotations)
+  }
+  return false
+}
+
 const separateChangeByFiles = async (
   change: DetailedChange,
   source: NaclFilesSource
 ): Promise<DetailedChange[]> => {
-  const isEmptyChangeElement = isEmptyElement(getChangeElement(change))
+  const isEmptyChangeElm = isEmptyChangeElement(getChangeElement(change))
   return (await Promise.all(
     (await source.getSourceRanges(change.id))
       .map(range => range.filename)
@@ -62,7 +81,7 @@ const separateChangeByFiles = async (
           change,
           changeData => filterByFile(change.id, changeData, fileElements),
         )
-        if (!isEmptyChangeElement && isEmptyElement(getChangeElement(filteredChange))) {
+        if (!isEmptyChangeElm && isEmptyChangeElement(getChangeElement(filteredChange))) {
           return undefined
         }
         return { ...filteredChange, path: toPathHint(filename) }
